requirements.txt
Flask
flask-cors
pymongo
python-dotenv
bcrypt
flask-session